requirements.txt.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
pillow==10.4.0
streamlit==1.32.2
opencv-python==4.9.0.80
matplotlib==3.8.4
scikit-learn==1.4.2
